Move the file that contains the function to get the replacement variables out of helpers folder
It's moved to the same level of other functions that are used to retrieve the data used for creating the seo integration
